# server/src/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Shoreline Care OS API",
    "version": "6.2.0",
    "description": "Open-source dietary, clinical nutrition, and food operations platform API for care facilities. Generated from the Express route definitions by server/scripts/generate-openapi.mjs — do not edit by hand; run `npm run docs:generate` in server/."
  },
  "servers": [
    {
      "url": "http://localhost:3001/api",
      "description": "Local Facility Gateway"
    },
    {
      "url": "https://api.shorelineops.com/api",
      "description": "Enterprise Cloud Sync"
    }
  ],
...
  "paths": {
...
    "/api/hardware/haccp/equipment": {
      "get": {
        "operationId": "get_hardware_haccp_equipment",
        "summary": "List temperature-monitored equipment (active only by default).",
        "tags": [
          "hardware"
        ],
        "parameters": [
          {
            "name": "includeInactive",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Query parameter read by the handler."
          }
        ],
        "description": "List temperature-monitored equipment (active only by default).\nhardware.haccpRead: kitchen execution + readonly surveyors. No vendors.\n\nAuthentication: JWT Bearer token; capabilities: hardware.haccpRead (admits: admin, manager, dietitian, dietary, server, staff, readonly).",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden — insufficient role"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-shoreline-auth": {
          "required": true,
          "mechanism": "jwt-bearer",
          "exactRoles": [
            "admin",
            "manager",
            "dietitian",
            "dietary",
            "server",
            "staff",
            "readonly"
          ],
          "capabilities": [
            "hardware.haccpRead"
          ]
        }
      },
      "post": {
        "operationId": "post_hardware_haccp_equipment",
        "summary": "Register a piece of monitored equipment. Manager/admin only. Replay after the gate.",
        "tags": [
          "hardware"
        ],
        "description": "Register a piece of monitored equipment. Manager/admin only. Replay after the gate.\n\nAuthentication: JWT Bearer token; capabilities: hardware.haccpManage (admits: manager, admin).",
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad request — validation failed"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden — insufficient role"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-shoreline-auth": {
          "required": true,
          "mechanism": "jwt-bearer",
          "exactRoles": [
            "manager",
            "admin"
          ],
          "capabilities": [
            "hardware.haccpManage"
          ]
        }
      }
    },
    "/api/hardware/haccp/equipment/{id}": {
      "patch": {
        "operationId": "patch_hardware_haccp_equipment_id",
        "summary": "Update target temp, frequency, name, or active flag. Manager/admin only. Replay after the gate.",
        "tags": [
          "hardware"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "description": "Update target temp, frequency, name, or active flag. Manager/admin only. Replay after the gate.\n\nAuthentication: JWT Bearer token; capabilities: hardware.haccpManage (admits: manager, admin).",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request — validation failed"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden — insufficient role"
          },
          "404": {
            "description": "Not found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-shoreline-auth": {
          "required": true,
          "mechanism": "jwt-bearer",
          "exactRoles": [
            "manager",
            "admin"
          ],
          "capabilities": [
            "hardware.haccpManage"
          ]
        }
      }
    },
    "/api/hardware/haccp/log-temp": {
      "post": {
        "operationId": "post_hardware_haccp_log_temp",
        "summary": "Log one temperature check. Manual entry, or a REAL probe reading passed in",
        "tags": [
          "hardware"
        ],
        "description": "Log one temperature check. Manual entry, or a REAL probe reading passed in\nfrom the tablet's probe abstraction (source='probe' + probeDevice + the\ndevice's own reading). The server never fabricates a reading: tempF is a\nrequired measured value.\nCompliance is computed server-side and a violation (non-compliant) CANNOT\nbe persisted without correctiveAction text — enforced here AND again by\nthe client flow, so a bare LOGGED record can never be created.\nhardware.haccpLog: kitchen execution roles log temps (no frontdesk/\nreadonly/vendors). Replay after the gate.\n\nAuthentication: JWT Bearer token; capabilities: hardware.haccpLog (admits: admin, manager, dietitian, dietary, staff, server).",
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad request — validation failed"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden — insufficient role"
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Unprocessable entity"
          },
          "500": {
            "description": "Server error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-shoreline-auth": {
          "required": true,
          "mechanism": "jwt-bearer",
          "exactRoles": [
            "admin",
            "manager",
            "dietitian",
            "dietary",
            "staff",
            "server"
          ],
          "capabilities": [
            "hardware.haccpLog"
          ]
        }
      }
    },
    "/api/hardware/haccp/logs": {
      "get": {
        "operationId": "get_hardware_haccp_logs",
        "summary": "Query persisted temp logs. ?start=YYYY-MM-DD&end=YYYY-MM-DD",
        "tags": [
          "hardware"
        ],
        "parameters": [
          {
            "name": "checkType",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Query parameter read by the handler."
          },
          {
            "name": "end",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Query parameter read by the handler."
          },
          {
            "name": "equipmentId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Query parameter read by the handler."
          },
          {
            "name": "start",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Query parameter read by the handler."
          },
          {
            "name": "violationsOnly",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Query parameter read by the handler."
          }
        ],
        "description": "Query persisted temp logs. ?start=YYYY-MM-DD&end=YYYY-MM-DD\n&equipmentId=&violationsOnly=true&checkType=food|equipment\n\nAuthentication: JWT Bearer token; capabilities: hardware.haccpRead (admits: admin, manager, dietitian, dietary, server, staff, readonly).",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden — insufficient role"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-shoreline-auth": {
          "required": true,
          "mechanism": "jwt-bearer",
          "exactRoles": [
            "admin",
            "manager",
            "dietitian",
            "dietary",
            "server",
            "staff",
            "readonly"
          ],
          "capabilities": [
            "hardware.haccpRead"
          ]
        }
      }
    },
    "/api/hardware/haccp/logs/{id}": {
      "patch": {
        "operationId": "patch_hardware_haccp_logs_id",
        "summary": "Attach (or update) the corrective action on a logged violation. This is the",
        "tags": [
          "hardware"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "description": "Attach (or update) the corrective action on a logged violation. This is the\nonly way an OPEN violation closes — the corrective action text is the\nclosure requirement. Setting an empty value is rejected.\n\nAuthentication: JWT Bearer token; capabilities: hardware.haccpLog (admits: admin, manager, dietitian, dietary, staff, server).",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request — validation failed"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden — insufficient role"
          },
          "404": {
            "description": "Not found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-shoreline-auth": {
          "required": true,
          "mechanism": "jwt-bearer",
          "exactRoles": [
            "admin",
            "manager",
            "dietitian",
            "dietary",
            "staff",
            "server"
          ],
          "capabilities": [
            "hardware.haccpLog"
          ]
        }
      }
    },
    "/api/hardware/haccp/schedule": {
      "get": {
        "operationId": "get_hardware_haccp_schedule",
        "summary": "Equipment temp schedule with due/overdue surfacing. hardware.haccpRead.",
        "tags": [
          "hardware"
        ],
        "description": "Equipment temp schedule with due/overdue surfacing. hardware.haccpRead.\n\nAuthentication: JWT Bearer token; capabilities: hardware.haccpRead (admits: admin, manager, dietitian, dietary, server, staff, readonly).",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden — insufficient role"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-shoreline-auth": {
          "required": true,
          "mechanism": "jwt-bearer",
          "exactRoles": [
            "admin",
            "manager",
            "dietitian",
            "dietary",
            "server",
            "staff",
            "readonly"
          ],
          "capabilities": [
            "hardware.haccpRead"
          ]
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT issued by POST /api/auth/login. Role claim gates requireRole ranks."
      },
      "ehrWebhookSignature": {
        "type": "apiKey",
        "in": "header",
        "name": "X-EHR-Signature",
        "description": "HMAC-SHA256 of the raw body with EHR_WEBHOOK_SECRET."
      },
      "kioskApiSecret": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "Bearer <KIOSK_API_SECRET> for kiosk punch webhooks."
      },
      "stripeWebhookSignature": {
        "type": "apiKey",
        "in": "header",
        "name": "stripe-signature",
        "description": "Stripe webhook signature with STRIPE_WEBHOOK_SECRET."
      },
      "setupSecret": {
        "type": "apiKey",
        "in": "header",
        "name": "x-setup-secret",
        "description": "SETUP_BOOTSTRAP_SECRET for first-time facility initialization."
      }
    },
...
  },
...
}