# server/src/routes/hardware.ts
/**
 * Hardware Routes — Shoreline v6.0
 *
 * POST   /api/hardware/print/tray-card          DISABLED (503 — see below)
 *
 * B13: simulated printer registry (GET /printers), simulated BLE probe scan
 * (GET /probes), simulated probe reads (GET /probes/:probeId/temperature) and
 * simulated HACCP logging (POST /probes/:probeId/log-haccp) were CUT.
 */

import { Router, Request, Response, NextFunction } from 'express'
import { randomUUID } from 'crypto'
import { pool } from '../db/pool'
import type { AuthRequest } from '../middleware/requireAuth'
import { requireCapability } from '../middleware/permissions'
import { idempotencyMiddleware } from '../middleware/idempotency'
import { globalWebhookEmitter } from '../webhooks/emitter'
import type { HaccpTempViolationEvent } from '../webhooks/events'

export const hardwareRouter = Router()

// ── C01: Durable HACCP temperature logging (migration 023) ───────────────────
// Real, persisted temperature compliance. Manual entry + real probe readings
// only — no simulated/fake readings are ever generated here (B13 cut the
// simulated probe endpoints). Every temp check persists with the recording
// user + timestamp; nothing returns a fabricated LOGGED record.
// NPO/allergen checks live in the tray/card flows and are untouched here:
// temp checks never serve food, so they cannot bypass clinical gates.

// ── Thermal Printing ──────────────────────────────────────────────────────────

/**
 * POST /api/hardware/print/tray-card — DISABLED (review 2026-10-01).
 *
 * The legacy handler could not print safely: fluid consistency has no
 * canonical column (so it assumed a default), it printed NPO/inactive
 * residents as serviceable, and idempotent replay could re-issue an obsolete
 * clinical job. The endpoint stays mounted behind the hardware.print
 * capability gate and answers 503 with an explicit code — no database reads
 * or writes, no print job, no idempotency replay.
 *
 * Canonical path: GET /api/kitchen/traycards-generated, the validated signed
 * tray contract (kitchen execution/printing capability). Print from cards
 * generated there once this endpoint shares that contract.
 */
hardwareRouter.post('/print/tray-card', requireCapability('hardware.print'), (_req: Request, res: Response) => {
  return res.status(503).json({
    error: 'Legacy hardware tray-card printing is unavailable. Generate a signed tray card instead.',
    code: 'HARDWARE_PRINT_UNAVAILABLE',
    canonical: 'GET /api/kitchen/traycards-generated',
    guidance: 'Use GET /api/kitchen/traycards-generated (kitchen.write or hardware.print) for the validated signed tray contract; this legacy print path stays disabled until it shares that contract.',
  })
})


// ─────────────────────────────────────────────────────────────────────────────
// B13 scope cuts: the simulated hardware endpoints were CUT from this file:
//   GET    /api/hardware/printers                      (simulated printer registry)
//   GET    /api/hardware/probes                        (simulated BLE scan)
//   GET    /api/hardware/probes/:probeId/temperature   (simulated temp reads)
//   POST   /api/hardware/probes/:probeId/log-haccp     (simulated HACCP logging)
// Disabled (review 2026-10-01): POST /api/hardware/print/tray-card answers 503
// HARDWARE_PRINT_UNAVAILABLE until it shares the validated signed tray
// contract (GET /api/kitchen/traycards-generated).
// ─────────────────────────────────────────────────────────────────────────────

// ═════════════════════════════════════════════════════════════════════════════
// C01 — HACCP temperature compliance (durable, survey-evidence)
// ─────────────────────────────────────────────────────────────────────────────
// All writes are validated + audit-logged. Reads sit behind requireAuth
// (router-level) so the survey log never leaks to unauthenticated callers.
// ═════════════════════════════════════════════════════════════════════════════

const HACCP_EQUIPMENT_TYPES = ['fridge', 'freezer', 'dishwasher', 'hot-hold'] as const
const HACCP_CHECK_FREQUENCIES = ['shift', 'daily', 'weekly'] as const
// Equipment types that must stay COLD: compliant when measured <= target.
// All others (dishwasher final rinse, hot-hold, food cook/hold) are compliant
// when measured >= target.
const COLD_EQUIPMENT = new Set(['fridge', 'freezer'])
const FREQUENCY_HOURS: Record<string, number> = { shift: 8, daily: 24, weekly: 168 }

interface HaccpEquipmentRow {
  id: string
  name: string
  type: string
  target_temp_f: number
  check_frequency: string
  active: number | boolean
  created_at: string
  updated_at: string
}

async function auditHaccp(
  userId: string | null,
  action: string,
  resourceId: string,
  resourceType: string,
  details: Record<string, unknown>
): Promise<void> {
  await pool.query(
    `INSERT INTO audit_log (action, user_id, resource_id, resource_type, outcome, details)
     VALUES ($1, $2, $3, $4, 'success', $5)`,
    [action, userId, resourceId, resourceType, JSON.stringify(details)]
  )
}

async function getEquipment(id: string): Promise<HaccpEquipmentRow | null> {
  const { rows } = await pool.query(`SELECT * FROM haccp_equipment WHERE id = $1`, [id])
  return rows[0] ?? null
}

/** Due/overdue computation per equipment against its check frequency. */
async function buildSchedule(): Promise<Array<HaccpEquipmentRow & {
  last_log_id: string | null
  last_recorded_at: string | null
  last_temp_f: number | null
  next_due_at: string | null
  status: 'ok' | 'due' | 'overdue'
}>> {
  const { rows: equipment } = await pool.query(
    `SELECT * FROM haccp_equipment WHERE active = $1 ORDER BY name`, [true]
  )
  const now = Date.now()
  const out: Array<HaccpEquipmentRow & {
    last_log_id: string | null
    last_recorded_at: string | null
    last_temp_f: number | null
    next_due_at: string | null
    status: 'ok' | 'due' | 'overdue'
  }> = []
  for (const eq of equipment) {
    const { rows: last } = await pool.query(
      `SELECT id, recorded_at, temp_f FROM haccp_logs
       WHERE equipment_id = $1 AND check_type = 'equipment'
       ORDER BY recorded_at DESC LIMIT 1`,
      [eq.id]
    )
    const lastRow = last[0] ?? null
    const lastAt = lastRow ? new Date(lastRow.recorded_at).getTime() : null
    const intervalMs = (FREQUENCY_HOURS[eq.check_frequency] ?? 24) * 3600 * 1000
    let nextDueAt: string | null = null
    let status: 'ok' | 'due' | 'overdue'
    if (lastAt == null || Number.isNaN(lastAt)) {
      // Never logged → overdue (must be checked before it counts as covered).
      status = 'overdue'
    } else {
      const dueAt = lastAt + intervalMs
      nextDueAt = new Date(dueAt).toISOString()
      if (now >= dueAt) status = 'overdue'
      else if (now >= dueAt - 30 * 60 * 1000) status = 'due'
      else status = 'ok'
    }
    out.push({
      ...eq,
      last_log_id: lastRow?.id ?? null,
      last_recorded_at: lastRow?.recorded_at ?? null,
      last_temp_f: lastRow ? Number(lastRow.temp_f) : null,
      next_due_at: nextDueAt,
      status,
    })
  }
  return out
}

// ── GET /api/hardware/haccp/equipment ───────────────────────────────────────
// List temperature-monitored equipment (active only by default).
// hardware.haccpRead: kitchen execution + readonly surveyors. No vendors.
hardwareRouter.get('/haccp/equipment', requireCapability('hardware.haccpRead'), async (req: Request, res: Response, next: NextFunction) => {
  try {
    const includeInactive = req.query.includeInactive === 'true'
    const { rows } = await pool.query(
      `SELECT * FROM haccp_equipment ${includeInactive ? '' : 'WHERE active = $1'} ORDER BY name`,
      includeInactive ? [] : [true]
    )
    res.json(rows)
  } catch (err) { next(err) }
})

// ── POST /api/hardware/haccp/equipment ──────────────────────────────────────
// Register a piece of monitored equipment. Manager/admin only. Replay after the gate.
hardwareRouter.post('/haccp/equipment', requireCapability('hardware.haccpManage'), idempotencyMiddleware(), async (req: AuthRequest, res, next) => {
  try {
    const { name, type, targetTempF, checkFrequency = 'daily' } = req.body ?? {}
    if (typeof name !== 'string' || !name.trim()) {
      return res.status(400).json({ error: 'name is required' })
    }
    if (!(HACCP_EQUIPMENT_TYPES as readonly string[]).includes(type)) {
      return res.status(400).json({ error: `type must be one of ${HACCP_EQUIPMENT_TYPES.join(', ')}` })
    }
    const target = Number(targetTempF)
    if (!Number.isFinite(target)) {
      return res.status(400).json({ error: 'targetTempF must be a number (°F)' })
    }
    if (!(HACCP_CHECK_FREQUENCIES as readonly string[]).includes(checkFrequency)) {
      return res.status(400).json({ error: `checkFrequency must be one of ${HACCP_CHECK_FREQUENCIES.join(', ')}` })
    }
    const id = randomUUID()
    await pool.query(
      `INSERT INTO haccp_equipment (id, name, type, target_temp_f, check_frequency)
       VALUES ($1, $2, $3, $4, $5)`,
      [id, name.trim(), type, target, checkFrequency]
    )
    const saved = await getEquipment(id)
    await auditHaccp(req.userId ?? null, 'HACCP_EQUIPMENT_CREATE', id, 'haccp_equipment',
      { name: name.trim(), type, targetTempF: target, checkFrequency })
    res.status(201).json(saved)
  } catch (err) { next(err) }
})

// ── PATCH /api/hardware/haccp/equipment/:id ─────────────────────────────────
// Update target temp, frequency, name, or active flag. Manager/admin only. Replay after the gate.
hardwareRouter.patch('/haccp/equipment/:id', requireCapability('hardware.haccpManage'), idempotencyMiddleware(), async (req: AuthRequest, res, next) => {
  try {
    const existing = await getEquipment(req.params.id)
    if (!existing) return res.status(404).json({ error: 'equipment not found' })
    const { name, targetTempF, checkFrequency, active } = req.body ?? {}
    if (name !== undefined && (typeof name !== 'string' || !name.trim())) {
      return res.status(400).json({ error: 'name must be a non-empty string' })
    }
    const target = targetTempF !== undefined ? Number(targetTempF) : null
    if (targetTempF !== undefined && !Number.isFinite(target)) {
      return res.status(400).json({ error: 'targetTempF must be a number (°F)' })
    }
    if (checkFrequency !== undefined && !(HACCP_CHECK_FREQUENCIES as readonly string[]).includes(checkFrequency)) {
      return res.status(400).json({ error: `checkFrequency must be one of ${HACCP_CHECK_FREQUENCIES.join(', ')}` })
    }
    if (active !== undefined && typeof active !== 'boolean') {
      return res.status(400).json({ error: 'active must be a boolean' })
    }
    await pool.query(
      `UPDATE haccp_equipment
       SET name = COALESCE($2, name),
           target_temp_f = COALESCE($3, target_temp_f),
           check_frequency = COALESCE($4, check_frequency),
           active = COALESCE($5, active),
           updated_at = NOW()
       WHERE id = $1`,
      [req.params.id,
       name !== undefined ? name.trim() : null,
       target,
       checkFrequency ?? null,
       active ?? null]
    )
    const saved = await getEquipment(req.params.id)
    await auditHaccp(req.userId ?? null, 'HACCP_EQUIPMENT_UPDATE', req.params.id, 'haccp_equipment',
      { name: name ?? undefined, targetTempF: target ?? undefined, checkFrequency: checkFrequency ?? undefined, active: active ?? undefined })
    res.json(saved)
  } catch (err) { next(err) }
})

// ── GET /api/hardware/haccp/schedule ────────────────────────────────────────
// Equipment temp schedule with due/overdue surfacing. hardware.haccpRead.
hardwareRouter.get('/haccp/schedule', requireCapability('hardware.haccpRead'), async (_req: Request, res: Response, next: NextFunction) => {
  try {
    const schedule = await buildSchedule()
    const overdue = schedule.filter(s => s.status === 'overdue').length
    const due = schedule.filter(s => s.status === 'due').length
    res.json({ generatedAt: new Date().toISOString(), overdue, due, equipment: schedule })
  } catch (err) { next(err) }
})

// ── GET /api/hardware/haccp/logs ───────────────────────────────────────────
// Query persisted temp logs. ?start=YYYY-MM-DD&end=YYYY-MM-DD
// &equipmentId=&violationsOnly=true&checkType=food|equipment
hardwareRouter.get('/haccp/logs', requireCapability('hardware.haccpRead'), async (req: Request, res: Response, next: NextFunction) => {
  try {
    const { start, end, equipmentId, violationsOnly, checkType } = req.query
    const conditions: string[] = []
    const params: any[] = []
    const push = (cond: string, val: any) => { params.push(val); conditions.push(cond.replace('?', `$${params.length}`)) }

    if (start) push(`date(l.recorded_at) >= date(?)`, start)
    if (end) push(`date(l.recorded_at) <= date(?)`, end)
    if (equipmentId) push(`l.equipment_id = ?`, equipmentId)
    if (violationsOnly === 'true') push(`l.compliant = ?`, false)
    if (checkType && ['food', 'equipment'].includes(checkType as string)) push(`l.check_type = ?`, checkType)

    // Portable boolean literal: pool.ts maps JS booleans to 1/0 on SQLite;
    // PostgreSQL compares natively against its BOOLEAN column.
    const where = conditions.length ? 'WHERE ' + conditions.join(' AND ') : ''
    const { rows } = await pool.query(
      `SELECT l.*, e.name AS equipment_name, u.name AS recorded_by_name
       FROM haccp_logs l
       LEFT JOIN haccp_equipment e ON e.id = l.equipment_id
       LEFT JOIN users u ON CAST(u.id AS TEXT) = l.recorded_by
       ${where}
       ORDER BY l.recorded_at DESC
       LIMIT 500`,
      params
    )
    res.json(rows)
  } catch (err) { next(err) }
})

// ── POST /api/hardware/haccp/log-temp ───────────────────────────────────────
// Log one temperature check. Manual entry, or a REAL probe reading passed in
// from the tablet's probe abstraction (source='probe' + probeDevice + the
// device's own reading). The server never fabricates a reading: tempF is a
// required measured value.
// Compliance is computed server-side and a violation (non-compliant) CANNOT
// be persisted without correctiveAction text — enforced here AND again by
// the client flow, so a bare LOGGED record can never be created.
// hardware.haccpLog: kitchen execution roles log temps (no frontdesk/
// readonly/vendors). Replay after the gate.
hardwareRouter.post('/haccp/log-temp', requireCapability('hardware.haccpLog'), idempotencyMiddleware(), async (req: AuthRequest, res, next) => {
  try {
    const {
      checkType, itemName = '', equipmentId, tempF, targetTempF,
      correctiveAction = '', source = 'manual', probeDevice, recordedBy,
    } = req.body ?? {}

    if (!['food', 'equipment'].includes(checkType)) {
      return res.status(400).json({ error: "checkType must be 'food' or 'equipment'" })
    }
    const measured = Number(tempF)
    if (!Number.isFinite(measured) || measured < -40 || measured > 250) {
      return res.status(400).json({ error: 'tempF must be a measured temperature between -40 and 250 °F' })
    }
    if (typeof itemName !== 'string' || itemName.length > 200) {
      return res.status(400).json({ error: 'itemName must be a string of at most 200 characters' })
    }
    if (!['manual', 'probe'].includes(source)) {
      return res.status(400).json({ error: "source must be 'manual' or 'probe'" })
    }
    if (source === 'probe' && (typeof probeDevice !== 'string' || !probeDevice.trim())) {
      // A probe-sourced reading without an identified device is unverifiable —
      // treat it as manual rather than laundering an anonymous reading.
      return res.status(400).json({ error: 'probeDevice (device name) is required when source is probe' })
    }
    if (typeof correctiveAction !== 'string' || correctiveAction.length > 2000) {
      return res.status(400).json({ error: 'correctiveAction must be a string of at most 2000 characters' })
    }

    // Resolve target temperature + equipment context.
    let equipment: HaccpEquipmentRow | null = null
    let target: number
    if (checkType === 'equipment') {
      if (!equipmentId) {
        return res.status(400).json({ error: 'equipmentId is required for equipment checks' })
      }
      equipment = await getEquipment(equipmentId)
      if (!equipment) return res.status(404).json({ error: 'equipment not found' })
      if (!equipment.active) {
        return res.status(400).json({ error: 'equipment is deactivated — reactivate it before logging' })
      }
      target = targetTempF !== undefined ? Number(targetTempF) : Number(equipment.target_temp_f)
    } else {
      // Food checks default to the cook standard (165 °F); hot-hold checks
      // pass their own target (e.g. 140).
      target = targetTempF !== undefined ? Number(targetTempF) : 165
      if (equipmentId) {
        equipment = await getEquipment(equipmentId)
        if (equipment && equipment.active) target = Number(equipment.target_temp_f)
      }
    }
    if (!Number.isFinite(target)) {
      return res.status(400).json({ error: 'targetTempF must be a number (°F)' })
    }

    // Server-side compliance evaluation.
    const isColdEquipment = equipment ? COLD_EQUIPMENT.has(equipment.type) : false
    const compliant = isColdEquipment ? measured <= target : measured >= target
    const action = correctiveAction.trim()
    if (!compliant && !action) {
      // Hard rule: a violation cannot be closed without corrective-action
      // text. The row is NOT persisted — the entry stays open until the
      // corrective action is supplied.
      return res.status(422).json({
        error: 'Violation: temperature is out of range. A corrective action is required before this log can be closed.',
        measuredTempF: measured,
        targetTempF: target,
        compliant: false,
        violationType: deriveViolationType(checkType, equipment, target),
      })
    }

    const id = randomUUID()
    const by = typeof recordedBy === 'string' && recordedBy.trim()
      ? recordedBy.trim()
      : (req.userId ?? null)
    const violationType = compliant ? null : deriveViolationType(checkType, equipment, target)

    await pool.query(
      `INSERT INTO haccp_logs
         (id, check_type, item_name, equipment_id, temp_f, target_temp_f,
          compliant, violation_type, corrective_action, source, probe_device,
          recorded_by, recorded_at)
       VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10, $11, $12, NOW())`,
      [id, checkType, itemName.trim(), equipment?.id ?? null, measured, target,
       compliant, violationType, action,
       source, source === 'probe' ? probeDevice.trim() : null,
       by]
    )
    const { rows: savedRows } = await pool.query(
      `SELECT l.*, e.name AS equipment_name FROM haccp_logs l
       LEFT JOIN haccp_equipment e ON e.id = l.equipment_id
       WHERE l.id = $1`, [id]
    )
    const record = savedRows[0]
    if (!record) {
      return res.status(500).json({ error: 'HACCP log was not persisted' })
    }

    await auditHaccp(req.userId ?? null, 'HACCP_LOG_TEMP', id, 'haccp_log',
      { checkType, itemName: itemName.trim(), equipmentId: equipment?.id ?? null, tempF: measured, targetTempF: target, compliant, violationType })

    // Fire the existing violation webhook on real violations (subscribers
    // already exist for haccp.temp.violation — B13 only cut the simulated
    // producer, not the event type).
    if (!compliant) {
      const event: HaccpTempViolationEvent = {
        event: 'haccp.temp.violation',
        eventId: randomUUID(),
        facilityId: process.env.FACILITY_ID ?? 'FAC-001',
        emittedAt: new Date().toISOString(),
        payload: {
          probeId: source === 'probe' ? (probeDevice as string) : 'manual-entry',
          stationId: equipment ? equipment.name : 'food-check',
          itemName: itemName.trim() || (equipment ? equipment.name : 'food item'),
          measuredTempF: measured,
          requiredMinTempF: target,
          violationType: violationType as HaccpTempViolationEvent['payload']['violationType'],
          loggedBy: typeof by === 'string' ? by : 'unknown',
          correctionRequired: true,
        },
      }
      globalWebhookEmitter.emit(event).catch((err) => {
        console.error('[HACCP] Webhook emit error:', err)
      })
    }

    res.status(201).json({ success: true, record })
  } catch (err) { next(err) }
})

// ── PATCH /api/hardware/haccp/logs/:id ─────────────────────────────────────
// Attach (or update) the corrective action on a logged violation. This is the
// only way an OPEN violation closes — the corrective action text is the
// closure requirement. Setting an empty value is rejected.
hardwareRouter.patch('/haccp/logs/:id', requireCapability('hardware.haccpLog'), idempotencyMiddleware(), async (req: AuthRequest, res, next) => {
  try {
    const { correctiveAction } = req.body ?? {}
    if (typeof correctiveAction !== 'string' || !correctiveAction.trim()) {
      return res.status(400).json({ error: 'correctiveAction text is required — a violation cannot be closed without it' })
    }
    if (correctiveAction.length > 2000) {
      return res.status(400).json({ error: 'correctiveAction must be at most 2000 characters' })
    }
    const { rows } = await pool.query(`SELECT * FROM haccp_logs WHERE id = $1`, [req.params.id])
    if (!rows[0]) return res.status(404).json({ error: 'log not found' })
    await pool.query(
      `UPDATE haccp_logs SET corrective_action = $1 WHERE id = $2`,
      [correctiveAction.trim(), req.params.id]
    )
    const { rows: saved } = await pool.query(`SELECT * FROM haccp_logs WHERE id = $1`, [req.params.id])
    await auditHaccp(req.userId ?? null, 'HACCP_CORRECTIVE_ACTION', req.params.id, 'haccp_log',
      { correctiveAction: correctiveAction.trim() })
    res.json(saved[0])
  } catch (err) { next(err) }
})

function deriveViolationType(
  checkType: string,
  equipment: HaccpEquipmentRow | null,
  target: number
): 'hot_hold_below_140' | 'cold_hold_above_41' | 'cook_below_165' {
  if (equipment && COLD_EQUIPMENT.has(equipment.type)) return 'cold_hold_above_41'
  if (checkType === 'food' && target >= 165) return 'cook_below_165'
  return 'hot_hold_below_140'
}